+++ vllm/entrypoints/serve/__init__.py
@@ -30,3 +30,32 @@ def register_vllm_serve_api_routers(app: FastAPI):
     )
 
     attach_tokenize_router(app)
+
+
+def register_vllm_dev_mode_api_routers(app: FastAPI):
+    logger.warning(
+        "SECURITY WARNING: Development endpoints are enabled! "
+        "This should NOT be used in production!"
+    )
+
+    from .dev_mode.cache.api_router import attach_router as attach_cache_router
+
+    attach_cache_router(app)
+
+    from .dev_mode.rlhf.api_router import attach_router as attach_rlhf_router
+
+    attach_rlhf_router(app)
+
+    from .dev_mode.rpc.api_router import attach_router as attach_rpc_router
+
+    attach_rpc_router(app)
+
+    from .dev_mode.server_info.api_router import (
+        attach_router as attach_server_info_router,
+    )
+
+    attach_server_info_router(app)
+
+    from .dev_mode.sleep.api_router import attach_router as attach_sleep_router
+
+    attach_sleep_router(app)
